[fix] Fix issue with --split-chapters option.
When using the --split-chapters option of yt-dlp, the splited files are not saved in the specified location. To fix that, I suggest to use -P (path) instead of -o (output format) in the default args. This don't change the behaviour for single files, as yt-dlp apply something like "%(title)s.%(ext)s" by default for output file name. That way, the user can "override" (define) the -o option in his preset to have better control of the output file (or fileS if --split-chapters)